requirements.txt
Flask==3.1.0
Flask-Login==0.6.3
Flask-Session==0.8.0
Flask-WTF==1.2.1
WTForms==3.2.1
bcrypt==4.2.1
python-dotenv==1.0.1
pandas==2.2.3
openpyxl==3.1.5
reportlab==4.2.5
matplotlib==3.9.3
Flask-Mail==0.9.1
itsdangerous==2.2
email-validator==2.2.0
Flask-Limiter==3.5.1
Flask-Talisman==1.1.0
pyodbc==5.0.1
